+++ rust/src/api/mod.rs
@@ -3,7 +3,8 @@ pub mod secret;
 pub mod simple;
 pub mod sync;
 pub mod voting;
-pub mod voting_config;
 pub mod wallet;
 
 mod voting_helpers;
+
+pub use crate::api::voting as voting_config;
